--- dev/srcsearch/format.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"encoding/json"
-)
-
-// json.MarshalIndent, but with defaults.
-func marshalIndent(v interface{}) ([]byte, error) {
-	return json.MarshalIndent(v, "", "  ")
-}
